Improve sidebar
