+++ gui_dev/package.json
@@ -16,6 +16,7 @@
     "immer": "^10.1.1",
     "plotly.js-basic-dist-min": "^2.35.2",
     "react": "next",
+    "react-icons": "^5.2.1",
     "react-dom": "next",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
