--- gui_dev/src/components/Drawer/Drawer.module.css
@@ -1,8 +1,0 @@
-.drawer {
-  width: 300px;
-  height: 100%;
-  background-color: #ffffff;
-  box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
-  overflow-y: auto;
-  padding: 20px;
-}
