+++ gui_dev/src/components/TitleBar/TitleBar.module.css
@@ -15,7 +15,14 @@
   }
 }
 
-.buttons {
+.homeButton {
+  svg {
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.windowButtons {
   display: flex;
   flex-direction: row;
   justify-content: center;
