+++ gui_dev/src/pages/Welcome.jsx
@@ -1,0 +1,13 @@
+import styles from "./Welcome.module.css";
+import { useAppInfoStore } from "@/stores";
+
+export const Welcome = () => (
+  <div className={styles.welcomeContainer}>
+    <h2>PyNeuromodulation</h2>
+    <p className={styles.loading}>Loading application...</p>
+    <div className={styles.info}>
+      <p>Authors: John Doe, Jane Smith</p>
+      <p>License: MIT</p>
+    </div>
+  </div>
+);
