+++ gui_dev/vite.config.js
@@ -2,7 +2,6 @@ import { defineConfig } from "vite";
 import { fileURLToPath } from "url";
 import react from "@vitejs/plugin-react";
 import path from "path";
-import { fileURLToPath } from "url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
